scaleResolutionDownTo WPTs: Add simulcast and invalid resolution tests.
Simulcast was the last major piece of testing missing. With this CL we
should be good to go for shipping this API in a separate CL.

In addition to the new tests, this CL also tweaks some resolutions used
by other tests to avoid odd resolutions. Some encoder implementations
have resolutions alignments (e.g. 15p would become 16p). While not a
problem on Chrome which uses VP8 + libvpx by default, I predict this
would become a problem when the WPTs run on Safari having H264 by
default which would likely trigger resolution alignments so it's best
to use even resolutions when we have no reason not to.

Bug: chromium:363544347
Change-Id: Id1dba2968f3a0ad3e3206d15b117fdd18643ff41